Add docstrings and create separate taxonomy_classifier getters file
