+++ discovery_child_development/pipeline/models/taxonomy_classifier/02a_sentence_embeddings.py
@@ -17,7 +17,7 @@ from sentence_transformers import SentenceTransformer
 from time import time
 
 from discovery_child_development import logging, S3_BUCKET, config, taxonomy_config
-from discovery_child_development.getters import taxonomy
+from discovery_child_development.getters import taxonomy_classifier
 
 model = SentenceTransformer(taxonomy_config["sentence_embeddings_model"])
 
@@ -32,7 +32,7 @@ def embed_texts(
     vectors_out_path=VECTORS_OUT_PATH,
     vectors_filename=VECTORS_FILENAME,
 ):
-    labelled_text_df = taxonomy.get_training_data(split=split)
+    labelled_text_df = taxonomy_classifier.get_training_data(split=split)
 
     labelled_docs = labelled_text_df["text"].tolist()
     labelled_ids = labelled_text_df["id"].tolist()
